+++ mmc/trunk/src/simpmesh.h
@@ -120,7 +120,7 @@ extern "C" {
     void mesh_savedetphoton(float *ppath, void *seeds, int count, int seedbyte, mcconfig *cfg);
     void mesh_getdetimage(float *detmap, float *ppath, int count, mcconfig *cfg, tetmesh *mesh);
     void mesh_savedetimage(float *detmap, mcconfig *cfg);
-    float mesh_getdetidx(int photonid, int colcount, float* ppath, mcconfig* cfg);
+    uint32_t mesh_getdetidx(int photonid, int colcount, float* ppath, mcconfig* cfg);
     void mesh_srcdetelem(tetmesh *mesh,mcconfig *cfg);
     
     void tracer_init(raytracer *tracer,tetmesh *mesh,char methodid);
